+++ src/otomi-stack.ts
@@ -23,7 +23,7 @@ import {
   getTeamServicesJsonPath,
   objectToArray,
 } from './utils'
-import cloneRepo, { Repo } from './repo'
+import cloneRepo, { encrypt, Repo } from './repo'
 import {
   cleanEnv,
   CORE_VERSION,
@@ -85,23 +85,21 @@ export default class OtomiStack {
     this.loadValues()
   }
 
-  getSubSetting(type, key) {
+  getSetting(type, key) {
     return { [key]: this.db.db.get([type, key]).value() }
   }
 
-  setSubSetting(type, data, key) {
-    if (!isEmpty(data)) {
-      const ret = this.db.db
-        .get([type, key])
-        // @ts-ignore
-        .assign(data[key])
-        .write()
-      this.db.dirty = true
-      return ret
+  setSetting(type, data, key) {
+    if (isEmpty(data) || isEmpty(data[key])) {
+      throw new Error('Received empty payload...')
     }
-    // If it returns the same object, unchanged, then you'll know that there is no successful PUT,
-    // at least it will not write empty values.
-    return this.db.db.get([type, key]).value()
+    const ret = this.db.db
+      .get([type, key])
+      // @ts-ignore
+      .assign(data[key])
+      .write()
+    this.db.dirty = true
+    return ret
   }
 
   getSettings(): Settings {
@@ -242,6 +240,7 @@ export default class OtomiStack {
   async triggerDeployment(email: string): Promise<void> {
     console.log('DISABLE_SYNC: ', env.DISABLE_SYNC)
     this.saveValues()
+    encrypt()
 
     if (!env.DISABLE_SYNC) {
       await this.repo.save(email)
@@ -290,6 +289,7 @@ export default class OtomiStack {
     }
     const options = {
       users: [user],
+      clusters: [cluster],
       contexts: [context],
       currentContext: context.name,
     }
@@ -324,6 +324,7 @@ export default class OtomiStack {
 
   loadValues(skipAdmin = false): void {
     this.loadCluster()
+    this.loadPolicies()
     this.loadSettings()
     this.loadTeams(skipAdmin)
     this.db.setDirtyActive()
@@ -358,12 +359,19 @@ export default class OtomiStack {
     this.repo.writeFile(dataPath, plainData)
   }
 
+  loadPolicies(): void {
+    const data = this.loadConfig('./env/policies.yaml', './env/policies.yaml') as Settings
+    // @ts-ignore
+    this.db.db.get('settings').assign(data).write()
+  }
+
   loadSettings(): void {
     const data = this.loadConfig(
       './env/settings.yaml',
       `./env/secrets.settings.yaml${this.decryptedFilePostfix}`,
     ) as Settings
-    this.db.db.set('settings', data).write()
+    // @ts-ignore
+    this.db.db.get('settings').assign(data).write()
   }
 
   loadTeamJobs(teamId: string): void {
@@ -392,10 +400,7 @@ export default class OtomiStack {
     const secrets: Array<Secret> = get(data, getTeamSecretsJsonPath(teamId), [])
 
     secrets.forEach((inSecret) => {
-      // @ts-ignore
-      const secret: Secret = this.convertSecretToDb(inSecret, teamId)
-      const res: any = this.db.populateItem('secrets', secret, { teamId, name: secret.name }, secret.id)
-      console.log(`Loaded secret: name: ${res.name}, id: ${res.id}, teamId: ${teamId}`)
+      this.loadSecret(inSecret, teamId)
     })
   }
 
@@ -403,7 +408,7 @@ export default class OtomiStack {
     const mergedData: Core = this.loadConfig('./env/teams.yaml', `./env/secrets.teams.yaml${this.decryptedFilePostfix}`)
 
     Object.values(mergedData.teamConfig.teams).forEach((team: Team) => {
-      this.db.populateItem('teams', { ...team, name: team.id! }, undefined, team.id)
+      this.loadTeam(team)
       if (!skipAdmin) this.loadCoreServices()
       this.loadTeamJobs(team.id!)
       this.loadTeamServices(team.id!)
@@ -414,7 +419,7 @@ export default class OtomiStack {
   loadCoreServices(): void {
     const { services } = this.coreValues
     services.forEach((svc) => {
-      this.convertServiceToDb(svc, 'admin')
+      this.loadService(svc, 'admin')
     })
   }
 
@@ -427,13 +432,22 @@ export default class OtomiStack {
     const data = this.repo.readFile(relativePath)
     const services = get(data, getTeamServicesJsonPath(teamId), [])
     services.forEach((svc) => {
-      this.convertServiceToDb(svc, teamId)
+      this.loadService(svc, teamId)
     })
+  }
+
+  savePolicies(): void {
+    const settings: Settings = this.getSettings()
+    this.repo.writeFile('./env/policies.yaml', { policies: settings.policies })
   }
 
   saveSettings(): void {
     const settings: Settings = this.getSettings()
-    this.saveConfig('./env/settings.yaml', `./env/secrets.settings.yaml${this.decryptedFilePostfix}`, settings)
+    this.saveConfig(
+      './env/settings.yaml',
+      `./env/secrets.settings.yaml${this.decryptedFilePostfix}`,
+      omit(settings, 'policies'),
+    )
   }
 
   saveTeams(): void {
@@ -450,23 +464,20 @@ export default class OtomiStack {
     ]
     const secretPaths: string[] = []
     const teams = this.getTeams()
-    teams.forEach((team) => {
-      // TODO: fix this ugly team.id || ''
-      this.saveTeamJobs(team.id || '')
-      this.saveTeamServices(team.id || '')
-      this.saveTeamSecrets(team.id || '')
-      // eslint-disable-next-line no-param-reassign
+    teams.forEach((inTeam) => {
+      const team: any = omit(inTeam, 'name')
+      this.saveTeamJobs(team.id!)
+      this.saveTeamServices(team.id!)
+      this.saveTeamSecrets(team.id!)
       if (!team.password) team.password = generatePassword(16, false)
-      teamValues[team.id || ''] = omit(team, 'name')
+      team.resourceQuota = arrayToObject(team.resourceQuota ?? [])
 
       secretPropertyPaths.forEach((propertyPath) => {
         secretPaths.push(`teamConfig.teams.${team.id}.${propertyPath}`)
       })
+      teamValues[team.id!] = team
     })
-
-    const values = {}
-    set(values, 'teamConfig.teams', teamValues)
-
+    const values = set({}, 'teamConfig.teams', teamValues)
     this.saveConfig(filePath, secretFilePath, values, secretPaths)
   }
 
@@ -480,7 +491,6 @@ export default class OtomiStack {
     const jobs = this.db.getCollection('jobs', { teamId })
     const jobsRaw = jobs.map((item) => omit(item, ['teamId']))
     const data = {}
-
     set(data, getTeamJobsJsonPath(teamId), jobsRaw)
     const filePath = getTeamJobsFilePath(teamId)
     this.repo.writeFile(filePath, data)
@@ -500,17 +510,25 @@ export default class OtomiStack {
     this.repo.writeFile(filePath, data)
   }
 
-  convertSecretToDb(inSecret, teamId): void {
+  loadTeam(inTeam): any {
+    const team = { ...inTeam }
+    team.resourceQuota = objectToArray(inTeam.resourceQuota)
+    const res: any = this.db.populateItem('teams', { ...team, name: team.id! }, undefined, team.id)
+    console.log(`Loaded team: ${res.name}, id: ${res.id}`)
+  }
+
+  loadSecret(inSecret, teamId): void {
     const secret: any = omit(inSecret, ...secretTransferProps)
     secret.teamId = teamId
     secret.secret = secretTransferProps.reduce((memo: any, prop) => {
       if (inSecret[prop] !== undefined) memo[prop] = inSecret[prop]
       return memo
     }, {})
-    return secret
+    const res: any = this.db.populateItem('secrets', secret, { teamId, name: secret.name }, secret.id)
+    console.log(`Loaded secret: name: ${res.name}, id: ${res.id}, teamId: ${teamId}`)
   }
 
-  convertDbSecretToValues(inSecret: any): void {
+  convertDbSecretToValues(inSecret: any): any {
     const secret: any = omit(inSecret, 'secret')
     secretTransferProps.forEach((prop) => {
       if (inSecret.secret[prop] !== undefined) secret[prop] = inSecret.secret[prop]
@@ -518,7 +536,7 @@ export default class OtomiStack {
     return secret
   }
 
-  convertServiceToDb(svcRaw, teamId): void {
+  loadService(svcRaw, teamId): void {
     // Create service
     const svc = omit(svcRaw, 'domain', 'forwardPath', 'hasCert', 'auth', 'ksvc', 'paths', 'type', 'ownHost', 'tlsPass')
     svc.teamId = teamId
@@ -531,7 +549,7 @@ export default class OtomiStack {
       } else {
         svc.ksvc = cloneDeep(svcRaw.ksvc)
         svc.ksvc.serviceType = 'ksvc'
-        svc.ksvc.annotations = objectToArray(svcRaw.ksvc.annotations, 'name', 'value')
+        svc.ksvc.annotations = objectToArray(svcRaw.ksvc.annotations)
         svc.ksvc.env = objectToArray(svcRaw.ksvc.env, 'name', 'value')
         svc.ksvc.files = objectToArray(svcRaw.ksvc.files, 'path', 'content')
         svc.ksvc.secretMounts = objectToArray(svcRaw.ksvc.secretMounts, 'name', 'path')
@@ -574,8 +592,8 @@ export default class OtomiStack {
     delete ksvc.serviceType
     if (serviceType === 'ksvc') {
       svcCloned.ksvc = ksvc
-      svcCloned.ksvc.annotations = arrayToObject(svc.ksvc.annotations ?? [], 'name', 'value')
-      svcCloned.ksvc.env = arrayToObject(svc.ksvc.env ?? [], 'name', 'value')
+      svcCloned.ksvc.annotations = arrayToObject(svc.ksvc.annotations!)
+      svcCloned.ksvc.env = arrayToObject(svc.ksvc.env ?? [])
       svcCloned.ksvc.files = arrayToObject(svc.ksvc.files ?? [], 'path', 'content')
       svcCloned.ksvc.secretMounts = arrayToObject(svc.ksvc.secretMounts ?? [], 'name', 'path')
       svcCloned.ksvc.command = svc.ksvc.command?.length > 1 ? svc.ksvc.command.split(' ') : svc.ksvc.command
@@ -606,6 +624,7 @@ export default class OtomiStack {
 
   saveValues(): void {
     // TODO: saveApps()
+    this.savePolicies()
     this.saveSettings()
     this.saveTeams()
   }