+++ src/repo.ts
@@ -16,14 +16,16 @@ const baseUrl = `http://${env.TOOLS_HOST}:17771/`
 const decryptUrl = `${baseUrl}decrypt`
 const encryptUrl = `${baseUrl}encrypt`
 
-async function decrypt(): Promise<AxiosResponse | void> {
+export async function decrypt(): Promise<AxiosResponse | void> {
   if (!env.USE_SOPS) return Promise.resolve()
+  console.info('Requesting decrypt action')
   const res = await axios.get(decryptUrl)
   return res
 }
 
-async function encrypt(): Promise<AxiosResponse | void> {
+export async function encrypt(): Promise<AxiosResponse | void> {
   if (!env.USE_SOPS) return Promise.resolve()
+  console.info('Requesting encrypt action')
   const res = await axios.get(encryptUrl)
   return res
 }
@@ -79,7 +81,7 @@ export class Repo {
     const absolutePath = path.join(this.path, relativePath)
     console.debug(`Writing to file: ${absolutePath}`)
     const cleanedData = cleanDeep(data)
-    const yamlStr = yaml.safeDump(cleanedData)
+    const yamlStr = yaml.safeDump(cleanedData, { indent: 4 })
     fs.writeFileSync(absolutePath, yamlStr, 'utf8')
   }
 
@@ -96,7 +98,6 @@ export class Repo {
   }
 
   async commit(author: string): Promise<CommitResult> {
-    await encrypt()
     await this.git.add('./*')
     const commitSummary = await this.git.commit(`otomi-api<${author}>`)
     console.debug(`Commit summary: ${JSON.stringify(commitSummary)}`)
@@ -110,15 +111,17 @@ export class Repo {
     if (!isRepo) {
       console.info(`Repo does not exist. Cloning from: ${this.url} to: ${this.path}`)
       await this.git.clone(this.repoPathAuth, this.path)
-      await decrypt()
       return
     }
     console.log('Repo already exists. Checking out correct branch.')
     await this.git.checkout(this.branch)
 
+    if (env.isDev) await decrypt() // do it now because pull usually fails because of dirty state of git
     try {
       await this.pull()
+      if (!env.isDev) await decrypt()
     } catch (e) {
+      console.error(e)
       if (env.isDev) await this.git.clean(CleanOptions.FORCE)
       else throw e
     }
@@ -126,7 +129,6 @@ export class Repo {
 
   async pull(): Promise<any> {
     const pullSummary = await this.git.pull(this.remote, this.branch, { '--rebase': 'true' })
-    await decrypt()
     console.debug(`Pull summary: ${JSON.stringify(pullSummary)}`)
     return pullSummary
   }